+++ src/cmd_nwr/mod.rs
@@ -2,7 +2,6 @@
 
 pub mod append;
 pub mod ardb;
-pub mod assembly;
 pub mod download;
 pub mod info;
 pub mod lineage;