+++ tests/cli_nwr.rs
@@ -226,7 +226,7 @@ fn command_append_rank() -> anyhow::Result<()> {
 fn command_assembly() -> anyhow::Result<()> {
     let mut cmd = Command::cargo_bin("nwr")?;
     let output = cmd
-        .arg("assembly")
+        .arg("template")
         .arg("tests/assembly/Trichoderma.assembly.tsv")
         .arg("-o")
         .arg("stdout")
@@ -237,28 +237,6 @@ fn command_assembly() -> anyhow::Result<()> {
 
     assert_eq!(stderr.lines().count(), 6);
     assert!(stderr.contains("Create url.tsv"));
-
-    assert!(stdout.lines().count() > 100);
-    assert!(stdout.contains("T_atrov"));
-
-    Ok(())
-}
-
-#[test]
-fn command_biosample() -> anyhow::Result<()> {
-    let mut cmd = Command::cargo_bin("nwr")?;
-    let output = cmd
-        .arg("biosample")
-        .arg("tests/assembly/Trichoderma.assembly.tsv")
-        .arg("-o")
-        .arg("stdout")
-        .output()
-        .unwrap();
-    let stdout = String::from_utf8(output.stdout).unwrap();
-    let stderr = String::from_utf8(output.stderr).unwrap();
-
-    assert_eq!(stderr.lines().count(), 3);
-    assert!(stderr.contains("Create sample.tsv"));
 
     assert!(stdout.lines().count() > 100);
     assert!(stdout.contains("T_atrov"));